change structure of utils
