+++ client/src/utils/index.js
@@ -1,0 +1,11 @@
+import getContract from './getContract'
+import web3 from './web3'
+import blockEvent from './blockEvent'
+import userIdentification from './userIdentification'
+
+export default {
+  getContract,
+  web3,
+  blockEvent,
+  userIdentification
+}
